Fix #54 Refactor folder structure and rename some files.
