+++ src/components/layout/hero.tsx
