+++ src/lib/shapes/data/starPath.ts
